Backend/requirements.txt
Flask
python-dotenv
Flask-RESTful
Flask-Restx
Flask-Cors
PyJWT
PyYAML
python-jose
passlib
pymongo
mongoengine